Include Npcap in Packetbeat when bundled with Agentbeat (#41909)
